backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-socketio==5.10.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
openai==1.3.0
google-generativeai==0.3.0
Werkzeug==3.0.1

